requirements.txt
Flask
mysql_connector_repackaged
Werkzeug
